requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-dotenv
sqlalchemy
python-jose
passlib[bcrypt]
pyodbc
sqlalchemy

